+++ Tests/SwifTumbTests/SwifTumbTests.swift
@@ -109,7 +109,7 @@ class SwifTumbTests: XCTestCase {
         
         let handle = try! self.client!.userDashboard(
             success: { (response: SwifTumbResponse) in
-                var post: SwifTumbResponse.Post
+                var post: SwifTumb.Post
                 post = (response.response?.posts?[0])!
                 XCTAssertEqual(200, response.meta.status, "blog posts")
                 let params = SwifTumb.PostReblogParameters(
